video_service/requirements.txt
fastapi>=0.100.0,<0.101.0
uvicorn>=0.15.0
python-multipart>=0.0.5
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
pydantic>=2.0.0
python-dotenv>=0.19.0
elevenlabs>=0.2.24
moviepy>=1.0.3
firebase-admin>=5.0.0
pytest>=6.2.5
pytest-asyncio>=0.16.0
httpx>=0.23.0
# Director dependencies will be installed from Director's requirements.txt
